+++ src/ts/domain/ProductManagementDomain/validator.ts
@@ -1,13 +1,12 @@
 import {
-  CASH_RULE,
   MAX_NAME_LENGTH,
   MAX_QUANTITY,
   MESSAGE,
   PRICE_RULE,
-} from '../constants';
-import { isInvalidNumber } from '../utils/validator';
-import Product from '../domain/Product';
-import { ProductInfo, ProductInfoUnionType } from '../domain/types';
+} from '../../constants';
+import { isInvalidNumber } from '../../utils/validator';
+import Product from '../Product';
+import { ProductInfo, ProductInfoUnionType } from '../types';
 
 const findEmptyField = (
   product: ProductInfo,
@@ -97,11 +96,4 @@ const validateProductInfo = (
   });
 };
 
-const validateCash = (cash: number) => {
-  if (isInvalidNumber(cash, CASH_RULE)) {
-    throw new Error(MESSAGE.ERROR_INVALID_CASH);
-  }
-  return true;
-};
-
-export { validateProductInfo, validateCash };
+export { validateProductInfo };
